AWS/AWS_app/aws_deployer_app/requirements.txt
# FastAPI and ASGI server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
python-multipart>=0.0.6

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Environment management
python-dotenv>=1.0.0

# SSH/Remote access (for legacy support, SSM is now primary)
paramiko>=3.4.0

# MongoDB
motor>=3.3.0
pymongo>=4.6.0

# Google OAuth
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
